Moved custom field type presentation into shared
ref https://linear.app/ghost/issue/BER-3846/

- how a field type looks in a picker (its icon, and the row pairing that icon
  with its name) is presentation shared by every surface that offers types, and
  there is about to be a second one in the members import
- Settings and the import had rendered the same markup independently, so the
  option row moves alongside the icon rather than being copied next to it
- no behaviour change: the option resolves its own label from the catalog, which
  is where the local {label, value} shape was reading it from anyway
